refactor: use correct imports in navigations
